Better Database Usage

+++ src/main/scala/cleanArch/domain/auth/Session.scala
@@ -1,8 +1,9 @@
 package cleanArch.domain.auth
 
-case class Session(userId: Int, isLogin: Boolean) {
+case class Session(userId: Long, isLogin: Boolean) {
 
   def updateState(state: Boolean): Session = {
     copy(isLogin = state)
   }
+
 }
